# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: VICIdial Agent UI API
  version: 1.1.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VicidialSettingsRequest:
      type: object
      required: [baseUrl, apiUser, apiPass, source]
      properties:
        baseUrl: { type: string }
        apiUser: { type: string }
        apiPass: { type: string }
        source: { type: string }
    VicidialSettingsResponse:
      type: object
      properties:
        baseUrl: { type: string }
        apiUser: { type: string }
        source: { type: string }
        updatedAt: { type: string, format: date-time }
...
    AgentPassUpdateRequest:
      type: object
      required: [agentPass]
      properties:
        agentPass: { type: string }
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
...
paths:
...
  /api/admin/settings:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VicidialSettingsResponse' }
    put:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VicidialSettingsRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
  /api/admin/users/{id}/agent-pass:
    put:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentPassUpdateRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
  /api/reports/summary:
    get:
      deprecated: true
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: campaign, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reports/export:
    get:
      deprecated: true
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: format, in: query, required: true, schema: { type: string, enum: [csv] } }
      responses:
        '200':
          description: csv
  /api/admin/summary:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/agents:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/campaigns:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/interactions:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaign, in: query, required: false, schema: { type: string } }
        - { name: agentUser, in: query, required: false, schema: { type: string } }
        - { name: dispo, in: query, required: false, schema: { type: string } }
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: size, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/interactions/export.csv:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaign, in: query, required: false, schema: { type: string } }
        - { name: agentUser, in: query, required: false, schema: { type: string } }
        - { name: dispo, in: query, required: false, schema: { type: string } }
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          description: csv
  /api/admin/users:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/users/{id}:
    put:
      deprecated: true
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...